% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextFind.R
\name{contextFind}
\alias{contextFind}
\title{Search R Files for Matching Text, Get Context and Links}
\usage{
contextFind(search_text, path = ".", recursive = TRUE, contextLines = 2)
}
\arguments{
\item{search_text}{Text to search for}

\item{path}{Root directory to search}

\item{recursive}{Whether to recurse into subfolders}

\item{contextLines}{Number of context lines around match}
}
\value{
A list of matches
}
\description{
Search R Files for Matching Text, Get Context and Links
}
